Fix clippy lints in airwolf
Make the cart clean under `cargo clippy -- -D warnings` so it passes the
new CI lint gate: use inspect_err over map_err, `if let` over `Option::map`
with a unit closure, drop redundant i16->i16 casts, and const NonZeroU8.

+++ airwolf/src/explosion.rs
@@ -61,8 +61,8 @@ impl Explosion {
                  radius,
                  ..
              }| {
-                let x = self.pos.x as i16 + *x_offset as i16;
-                let y = self.pos.y as i16 + *y_offset as i16;
+                let x = self.pos.x + *x_offset as i16;
+                let y = self.pos.y + *y_offset as i16;
                 let radius = *radius as f32 / u8::MAX as f32 * 2.0;
                 gfx.circle_fill(x, y, radius as u16, SPARK_COLOR);
             },
